requirement.txt
streamlit
langchain 
langchain_ollama
langchain_google_genai
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
pandas